+++ src/components/promo/team/index.tsx
@@ -6,7 +6,7 @@ import Title from "./title";
 function Team() {
   return (
     <CanvasComponent offset={coordinates.team}>
-      <div className="my-auto flex h-screen w-screen flex-col items-center justify-center">
+      <div className="flex h-screen w-screen scale-90 flex-col items-center justify-center">
         <Title />
         <Book />
       </div>
