+++ src/components/promo/team/title.tsx
@@ -1,13 +1,13 @@
 function Title() {
   return (
-    <div className="mb-4 flex flex-col gap-4 text-center">
+    <div className="-mt-8 mb-4 flex flex-col gap-4 text-center">
       <h1 className="my-2 font-jetbrains-mono text-lg font-medium text-medium">
         OUR TEAM
       </h1>
       <div className="font-dico text-3xl text-heavy">
         Meet the team behind the event!
       </div>
-      <div className="font-figtree text-medium">
+      <div className="mb-4 font-figtree text-medium">
         We&apos;re here to help bring your ideas to life.
       </div>
     </div>
